refactor: moved tests outside subfolder
